+++ package.json
@@ -35,5 +35,8 @@
     "react-uuid": "^2.0.0",
     "typescript": "5.0.2",
     "yup": "^1.0.2"
+  },
+  "devDependencies": {
+    "@types/papaparse": "^5.3.7"
   }
 }
