+++ src/sprintFiles/uploadHouseListBtn.tsx
@@ -1,0 +1,9 @@
+import Button from '@mui/material/Button';
+import AddIcon from '@mui/icons-material/Add';
+const uploadHouseListBtn = () => {
+    
+    return (
+        <Button>Add Members <AddIcon/></Button>
+    )
+}
+export default uploadHouseListBtn
